Create random_pass_generator
This program collects the user's first name and last name, welcomes the user, and creates random username and password details for the user. This could prove very useful for interfaces that require log-in. It saves time by randomly creating these data, and informing the user to change them later in the course of usage.